Update BPF testing mechanism for CI
Summary:
This switches to creating a GCE instance on demand with the proper image to pick the appropriate kernel version,
and then triggers the tests to run on docker in the given instance.
This updates the primary test trigger to still run from k8s and makes our BPF tests look the same as all other tests
at the topmost level.

Test Plan:
Run tests using the new methodolgy in Jenkins. Make sure that the tests run as expected, the runtime is stable and
testlogs are uploaded appropriately at the end.

Reviewers: zasgar, michelle

Reviewed By: zasgar

Subscribers: oazizi

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12622

GitOrigin-RevId: 13e13749e9f02282461c267f0b1a22bdb37a4b9e