[GH-593] Command Palette: Modal Dialog
Summary:
{F198974}

Moves the command palette into a modal dialog of its own.
Also adds a mention that this is a preview feature, with a direct link to file a GitHub issue.

The topbar now has a button to summon it instead of inlining it.

This allows way more space to operate, and a cleaner look.

Followup diffs will do the following:
- Make the run button in this modal functional
- Make Ctrl/Cmd+K open and close the modal from anywhere
- Make Ctrl/Cmd+Enter, while the modal is open, do the same thing as clicking the run button when active
- Switch the icons out for more applicable ones
- Clean up any remaining weird styling
- Future features to come after the holidays

Test Plan: Force the feature flag. Click the button in the topbar. Should have a spiffy command palette that doesn't actually run the commands it builds yet.

Reviewers: michelle, vihang

Reviewed By: michelle

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12620

GitOrigin-RevId: 148936ee82cf140262b9c1bcfbdfd9d5f67cdc3c