Fix vizier metadata that ended up in bad states
Summary:
Auto-repair was not done properly when a user didn't have a PVC. In those cases, the
operator would try to switch the user over to ETCD-backed metadata by setting useEtcdOperator
inside the vizier object. This wasn't fully fleshed out yet so when a user did this, they only
got the new metadata deployment, but not the new etcd deployment. Their cluster would now be
extra broken. Some users reported having a functioning metadata statefulset, but a failing
deployment.

This basically means there are a bunch of users with 2 different versions of metadata on their
cluster. This diff makes 1 of 2 fixes for those users
1. If the persistent metadata pod runs, operator switches to non-etcd mode in vizier.
2. If the persistent metadtaa pod does not run, operator switches to etcd mode in vizier.

The previous diff should make it so that switching to etcd mode should ensure that the system
works.

Test Plan:
- Faulty switch to etcd mode, when the auto-repair switched over from
  persistent to etcd mode.
  1. Create both etcd and persistent metadata deploys
  2. Force reconcile
    > Operator should kill the etcd deploy and switch to non-etcd mode in vizier
- Half-way switch to etcd mode
  1. Start from a working persistent version
  2. Scale down metadata statefulset to 0
  3. Delete pvc
  4. Scale up metadata statefulset to 1
  5. Deploy an etcd-metadata deploy w/o the etcd statefulset
  6. Force reconcile
      > Operator should kill the statefulset metadata, bring up etcd
      statefulset, and keep around the metadata deployment

Reviewers: vihang, michelle

Reviewed By: michelle

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12624

GitOrigin-RevId: 93167dba38d2ea4d6d8ea6968f22ed31982cc653