Always auto repair pending pvc-metadata
Summary:
We've tried a number of tactics to discover metadata installations that are incompatible
because of the pvc not binding. While this has worked to some degree, there are a number of gaps.

As an effort to get around many of the issues, this change tries to  pre-empt more
complicated checks by simply checking if the metadata pod is from a statefulset and is pending.

This approach is more broad and will likely generate some false positives, but during my cluster
investigation, those false positives are likley broken for some other reason so we won't cause
more problems than those that already exist.

Test Plan: Added unit tests and simulated a problem that we hope will be auto-repaired.

Reviewers: michelle, vihang, nserrino

Reviewed By: michelle

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12632

GitOrigin-RevId: f6b9b4abbe1eed7f2d8287035cd3579da0386c62