Add autopilot mode to Vizier crd
Summary:
Previous diffs were responsible for adding the vizier autopilot YAMLs/templates. This diff adds the autopilot field to the Vizier CRD so users can configure the mode as a flag.
Followup diffs will update the operator so that it can handle this flag properly.

Test Plan: This change just adds new fields which are unused, no change in functionality yet.

Reviewers: jamesbartlett, vihang, oazizi

Reviewed By: jamesbartlett

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12666

GitOrigin-RevId: 76a99dc71766913cd724090067fb6801ba47f0e8