Split lib/compiler build for llvm
Summary:
We llvm(clang) as both our compiler and part of the dependency for our system. This creates library tar files that we can import directly into our code base and not rely on the system directories.

This is not the final change needed for this. More cleanup and changes to the docker build process are coming. The chef provisioning is also not updated since we need to modify the bazel build first.

Test Plan: existing

Reviewers: vihang, jamesbartlett, #third_party_approvers

Reviewed By: vihang, #third_party_approvers

Signed-off-by: Zain Asgar <zasgar@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12700

GitOrigin-RevId: 5f35deb8936cc07b5c855a3a18145e8c02bb0029