Patch cpuinfo to support @platforms constraints.
Summary: Cpuinfo is an indirect dependency from tensorflow. It only supports the old style --cpu flags instead of the new platforms style. This patch supports the platforms constraints so that the arm cross build will work.

Test Plan: Tested with arm sysroots to see that the cpu flag doesn't matter.

Reviewers: vihang, zasgar, michelle, #third_party_approvers

Reviewed By: vihang, #third_party_approvers

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12771

GitOrigin-RevId: b32fc6c32d2b15d7ea5b39b1d5aeaca60aefcf49