[ci] Change ci/bazel_build_deps.sh to make sure only compatible targets are built.
Summary:
Changes ci/bazel_build_deps.sh to consider the bazel config and make sure only targets compatible with the targetted platform are built.
Without this change incompatible targets will be skipped with a warning but bazel will exit with an error code and message: "command succeeded, but not all targets were analyzed"

Test Plan: Tested with D12855, to make sure incompatible targets are not requested to be built.

Reviewers: vihang, zasgar, michelle

Reviewed By: zasgar

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12916

GitOrigin-RevId: c208fef130529b975682c81c746ee0c372e0cfd3