[ci] Fix cquery based compatibility check.
Summary: Fixes cquery platform compatibility check to skip targets that are in host or exec configs. These targets were causing the incompatibility stuff to not work, because cquery would view the host config as compatible so they'd still end up in the list. I also realized that the extra bazel query is unnecessary, as cquery supports the same syntax.

Test Plan: Tested with D12855, and saw that the incompatible targets don't show up in the list.

Reviewers: vihang, zasgar

Reviewed By: vihang

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12931

GitOrigin-RevId: fe118261921499ba7a6789134614f9454eab6df7