Ensure that genrules and other similar targets are built during CI
Summary:
The buildables rule was only filtering for binaries and libraries, However
we also depend on some genrules and other target types. This ensures
that we build those in CI too reducing the chances of breakages.

Test Plan:
Ran the bazel_build_deps script and ensured that
`//k8s/vizier:vizier_image_list` shows up in the list of targets.

Reviewers: zasgar, jamesbartlett

Reviewed By: jamesbartlett

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12928

GitOrigin-RevId: 6bf952f4530eed756b6eb568a5ef4edd7fbe16b9