In elf_reader_symbolizer_bpf_test, attach BPF probe to specific pid to fix runs_per_test.
Summary:
The elf reader symbolizer test uses a BPF probe to sample a stack trace of itself. If multiple instances of this test run in parallel, the probes may all attach to the same underlying memory location if the pid is not specificed when uprobe attach is invoked. Without specifying the pid, various instances of this test (running in parallel) can sample a stack trace from a different running instance. When it comes time to symbolize the vaddrs so sampled, the symbolizer cannot resolve the symbols because it assumes the symbols were from the self process rather than a peer process.

In this patch, we resolve the issue by attaching the probe to a specific pid, namely the self pid. With this patch, the vaddrs sampled by the stack trace sampler belong to the self process and the symbolizer correctly resolves symbols.

Test Plan: Ran with --runs_per_test=1024.

Reviewers: #stirling, ddelnano

Reviewed By: #stirling, ddelnano

Subscribers: ddelnano

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D12936

GitOrigin-RevId: 052f7187272699e1b0834599be06cead197d22f8