Simplify operator etcd repair (#792)
Summary: The operator tries to be smart by trying to switch the user to
an etcd-based deployment based on some conditions, as long as they have
been in that state for over 5 mins. These checks may be too hasty, as
sometimes a user's cluster may just be slower for various reasons. This
leads to some users getting unexpectedly moved to an etcd-based deploy.
There is also some extra logic to fix a state that users could have
gotten into because of auto-repair. Now that the operator has been out
in the wild for sometime, we can clean up this extra logic for
cleanliness/simplification.

A followup PR is coming to attempt to fix etcd-based deploys which may
have gotten to a bad state (lost quorum)

Relevant Issues: N/A

Type of change: /kind bug

Test Plan: Skaffold deploy operator

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>